Add files via upload
Simple Stack in C++ 
The MAX enum variable can be change to a const unsigned int 
then push{add item}
or pop{veiw item}

notice the 
++item in push
this is so that , when you push,  it should increment before adding the new item to a new position. 
the item-- 
in pop will show the item before moving to  a lower location. 

Lucky luck